reorganize the compressible Riemann solvers
